apply cleanup plugin

+++ src/main/java/muksihs/ipfs/photogallery/shared/Consts.java
@@ -2,7 +2,7 @@ package muksihs.ipfs.photogallery.shared;
 
 public class Consts {
 	public static final int KB = 1024;
-	public static final double thumbMaxSize = (int) 256;
+	public static final double thumbMaxSize = 256;
 	public static final double thumbJpgQuality = 0.70f;
 	public static final double imageMaxSize = (int) (2f*1024f);
 	public static final double imageJpgQuality = 0.92f;
